Share servicehost with vueHtml mode
