+++ server/src/embeddedSupport/languageModes.ts
@@ -116,7 +116,7 @@ export class LanguageModes {
     });
     const serviceHost = getServiceHost(tsModule, workspacePath, jsDocuments);
 
-    const vueHtmlMode = getVueHTMLMode(this.documentRegions, workspacePath, services.infoService);
+    const vueHtmlMode = getVueHTMLMode(serviceHost, this.documentRegions, workspacePath, services.infoService);
     const jsMode = await getJavascriptMode(
       serviceHost,
       this.documentRegions,
